Sped up initial creation of tailable cursor
Using a $natural sort on the `local.oplog.rs` capped collection is a significantly faster query than sorting on the unindexed `ts` field.
Also, specifying the oplogReplay allows us to act like a replica secondary, skipping forward through most of the oplog until our value of `ts` is found, i.e. not doing a full table scan to find the value of `ts`.